Fix name of binary in package file
The "old" name of this application prior to moving it
to GitHub was `git-alarms`. When getting it ready for
public distribution I changed the name but missed the
name of the binary used in package.json. This commit
fixes that so that the name of the binary matches
the documentation in the README.